reorganization of the build system
it was getting hairy to work with. I've renamed the folders containing the code for the subprojects nvcdrun and nvcdinfo
to their respective names (which were previously, test and util respectively - as we can see this was before a name that was reasonable had been derived).

this will make further maintenance much easier.

So, each subproject now has its own corresponding makefile. There is also a makefile, Makefile.inc, which contains the needed global variables
that apply to every makefile that's used to a build a target.

The interface remains the same as before: `make [DEBUG=1] libnvcd.so|nvcdrun|nvcdinfo` should be invoked at the root.
All binaries are still written to the toplevel bin directory.

Finally, nvcc is no longer used for linking or building libnvcd.so. libnvcd.so uses gcc with -std=gnu99, nvcdrun uses g++ with -std=c++11, and nvcdinfo uses the same.

For all cuda files, of course, nvcc is still used, and all of the main configuration flags present in the gcc/g++ are propagated over to nvcc (this was already the case, but now it's performed in a much nicer way)