[tests] test scripst for nvcdrun
A similar set of tests will be designed to be used with nvcdinfo.

All of these are meant primarily to test the respective programs
with interaction on supercomputers using LSF and JSM suites, but there is no
reason to limit them to only this usecase.

Note that the tests are generated by a configure script; see README
file that's included in this commit for details.