[nvcd_run_info] make num_runs static
this accomodates the recent simplification of the user initialization
process. before this change, the num_runs counter would always reset to
zero, since it was designed initially for repeated calls to be made in
between nvcd_host_begin/end.

since separate, individual kernels are going to be used, though, the
names of these kernels should be tracked.