[g_run_info] remove unique_ptr wrapper
if we don't, then when the application exits there will be a segfault.
we don't free the cupti memory here. Technically, we should, because if
we don't then the driver might not free it either, and this may be
memory that's leaking inside the OS or even possibly on the GPU.

there may be a better approach that can be leveraged, which I'm going to
look more into. It would involve using a static library instead of a
shared library, but I doubt anyone cares about that.