[tests] nvcdrun test improvements
All of these are fairly small scale tests. They're designed to test the interaction
between CUDA_VISIBLE_DEVICES and the current method of querying the GPU device handles
with CUDA in libnvcd.so.

JSM appears to handle the interaction between the driver and the GPU in terms of which GPU is mapped to which index.

For example, if we have 4 resource sets, each set with 1 GPU, querything the 0th GPU in libnvcd.so may provide a different
result.