[output] add event counter info
this implies also aggregating the counter data while printing the
difference from the previous aggregation (so we get only was recorded
for this iteration).

in the process of figuring out how to query a cuda kernel's name and the
metrics need to be listed in the output, so this is isn't completely
finished, but the most important information is there.