[ui] focus on brevity by not defaulting to "ALL" option for events and metrics
... when no value is specified for the respective variables. So, if the
environment variables are either blank (e.g. `BENCH_EVENTS=`) or not
defined at all, this is neither an error or a cause to do extra: simply
do nothing as a result of this.

Note that `BENCH_EVENTS=ALL` is still supported, but this behavior is
currently unavailable for `BENCH_METRICS` (and has been unavailable -
the only case where all metrics would be fetched is when the variable
itself was never defined).