[hook.cu] trivial bug fixes and restructuring.  [libnvcd.h] modifications to accomodate changes.
hook was restructed in temrs of how time is presented to the user, and also with some considerable assertion/safety checks.

libvcd includes a new pointer to function: `libnvcd_time_report()`, which is defined in the hook code and actually prints out the
recorded times.

also added documentation, some maintenance, etc.