[src] API improvements and cleanup with respect to moving toward multi-GPU
As it stands, `hook.cu` contains a direct copy of the code found in `nvcd.cuh`, which is currently a mess.

I'm gradually modifying this copied portion so that the transition to multi-GPU is smoother. This involves removing
globals and making adjustments to various other areas of the overall API (which includes the functions found declared in `cupti_util.h` and implemented in `cupti_util.c`)
to conform to these adjustments, while also modifying some code in `nvcd.cuh` here and there to avoid breaking compilation overall.

Eventually, what will remain within `nvcd.cuh` will be used for nvcdinfo and nothing more.