[hook] Rewrote trace client and timing functionality to allow for thread safety/independence and device independence.
In `commondef.h`, we also bump the error codes up to starting at 200 to reduce the likelihood of conflict, which beforehand was in conflict with Linux
error codes.

Note that a few modifications have been made to pre-existing code that removes global variable dependencies, such as
`g_time_records`, which stores each recording made by the timer.

There's still more work to do, such as reallocating the globals in `cupti_util.c` into a different area (either thread_local or the struct itself).

The original code for this file is mostly kept as it originally was; the implementation can be swapped out with the newer version for easier comparison.

That said, it will likely be removed at some point soon.