[cleanup1] remove no longer needed cuda utility header; fix formatting issues in nvcdinfo main source file; rename main source file.
Originally the cuda utility header, nvcd.cuh, was serving as more of a prototype.

All of the functionality is now essentially split between both nvcdinfo and the hook.

The functionality itself is mostly or less decoupled between the two, and so there is no reason to keep the nvcd.cuh header.

The main source file, originally nvcdinfo/src/main.cu, was a cuda source file in the beginning. Now, that also is no longer necessary.